test: first simple Integration Test for the PlayerController
PlayerController now has a simple implementation for the /create-user Mapping and sends a simple echo to the subscribers of /topic/create-user-response. Internally, the method saves the sent playerDto (sent as a JSON string) as a playerEntity to the database. It is yet to be tested if the saved entity can actually be retrieved  although since the PlayerEntityService.createPlayer method only calls the Repository's "save" method which is already tested in PlayerRepositoryTest.

fix: integration test works (Some legacy comments hindered even the demo tests from working. Note: with the current broker configuration, the demo integration tests don't work anymore)

fix: The ModelMapper Configuration was missing, so the modelMapper was not injectable. This is now fixed.